feat: add Contact object and Admin controller for viewing and deleting contacts
Added the Contact object and linked it to the front-end contact form. Also, added the Admin controller to enable the site administrators to view and delete the contact submissions. This feature enhances the site's functionality by enabling users to submit contact requests and providing a convenient way for site administrators to manage them.